app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django
django-admin-interface==0.29.1
django-colorfield==0.11.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==11.0.0
pip==23.0.1
python-slugify==8.0.4
requests==2.32.3
setuptools==65.5.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv
django-ckeditor
geoip2